Query current switches
An application can now query the Frenetic controller to get a list of
current switches and their ports. This is handy when switches are already
connected to the Frenetic-HTTP controller before the application starts.

I've updated topology discovery to use this feature and call "switch_up"
for every pre-existing switch. This should allow the controller to be
restarted.